// models/article/all-article-model.ts
export class AllArticleModel {

    creatorName?: string;
    articleId?:   string;
    slug?:        string;
    title?:       string;
    content?:     string;
    image?:       string;
    thumbnail?:   string;
    liked?:       number;
    shared?:      number;
    viewed?:      number;
    postDate?:    string;
    postDate2?:    string;

}